+++ routes/api.php
@@ -6,12 +6,14 @@ use App\Http\Controllers\Api\StudyDataController;
 use App\Http\Controllers\Api\TelegramWebhookController;
 use App\Http\Controllers\Api\SearchController;
 use App\Http\Controllers\Api\CalendarController;
+use App\Http\Controllers\Api\SupplementalTextController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/languages', [ReferenceDataController::class, 'languages']);
 Route::get('/translations', [ReferenceDataController::class, 'translations']);
 Route::get('/canons/{canon:code}/books', [ReferenceDataController::class, 'canonBooks']);
 Route::get('/translations/{translationCode}/books', [ReferenceDataController::class, 'translationBooks']);
+Route::get('/translations/{translationCode}/supplemental-texts', [SupplementalTextController::class, 'index']);
 Route::get('/translations/{translationCode}/books/{bookSlug}/chapters/{chapter}', [ChapterController::class, 'show'])
     ->whereNumber('chapter');
 Route::get('/strong/{number}', [StudyDataController::class, 'strongEntry']);